refactor the control and description
